backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-chess>=1.999
# python-oqs>=0.8.0  # Currently using mock implementation
pycryptodome>=3.17
Pillow>=9.0.0
liboqs-python